NA-698: meepctl ignores docker build failure and moved configKibana file
